Moved the export to form index so that the component jsx files could be named after what they are without an inconvenience of importing. Having component files named properly gives major convenience in that it makes it obvious as to what the VS Code tab is about.
