+++ src/components/Form/index.jsx
@@ -5,5 +5,15 @@ import SupportTable from '../SupportTable/SupportTable';
 import RadioButton from './RadioButtons/RadioButton';
 import SingleRadioButton from './RadioButton/RadioButton';
 import Address from './Address';
+import TextInput from './TextInput/TextInput';
 
-export { Button, Checkbox, Dropdown, SupportTable, RadioButton, Address, SingleRadioButton };
+export {
+    Button,
+    Checkbox,
+    Dropdown,
+    SupportTable,
+    RadioButton,
+    Address,
+    SingleRadioButton,
+    TextInput
+};
